requirements.txt
structurizr-python>=1.3
pandas>=2.0
openpyxl>=3.1